Update config.yml
Added a fan-out / fan-in workflow dependency for the Test and the Snyk security scan job to start the Publish job.  Note that there is a need to ensure that the dependency job(s) are able to run (watch out for branch filtering affecting this), else the Publish and Deploy jobs will not run at all.